refactor: move reader methods for indices to decoding
